still has errors don't merge! just check it out
