+++ lib/merchant_analyst_mod.rb
@@ -1,0 +1,15 @@
+module MerchantAnalyst
+
+  def merchants
+    @sales_engine.merchants
+  end
+
+  def all_merchants
+    merchants.merchants
+  end
+
+  def total_merchants
+    @sales_engine.merchants.all.count.to_f
+  end
+
+end
