+++ lib/sales_analyst.rb
@@ -1,6 +1,10 @@
 require_relative 'sales_engine'
+require_relative 'merchant_analyst_mod'
+require_relative 'item_analyst_mod'
 
 class SalesAnalyst
+  include MerchantAnalyst
+  include ItemAnalyst
 
   def initialize(sales_engine)
     @sales_engine = sales_engine
@@ -10,12 +14,8 @@ class SalesAnalyst
     (total_items / total_merchants).round(2)
   end
 
-  def total_merchants
-    @sales_engine.merchants.all.count.to_f
-  end
-
-  def total_items
-    @sales_engine.items.all.count.to_f
+  def mean_calculation_merchant(merchant)
+    (items.find_all_by_merchant_id(merchant.id).count - average_items_per_merchant) ** 2
   end
 
   def average_items_per_merchant_standard_deviation
@@ -26,31 +26,20 @@ class SalesAnalyst
   end
 
   def merchants_with_high_item_count
-    merchants.merchants.values.find_all do |merchant|
+    all_merchants.values.find_all do |merchant|
       merchant.items.count > (average_items_per_merchant_standard_deviation * 2)
     end
   end
 
-  def golden_items
-    total = items.all.sum do |item|
-      item.unit_price
-    end
-    mean = total / items.all.count
-    items.items.values.find_all do |item|
-      item.unit_price > (mean + average_item_price_standard_deviation * 2)
-    end
+  def average_item_price_standard_deviation
+    Math.sqrt(item_stdev).round(2)
   end
 
-  def average_item_price_standard_deviation
-    total = items.all.sum do |item|
-      item.unit_price
+  def golden_items
+    mean = all_items_price / total_items
+    all_items.values.find_all do |item|
+      item.unit_price > (mean + average_item_price_standard_deviation * 2)
     end
-    mean = total / items.all.count
-    variance = items.all.map do |item|
-      (item.unit_price - mean) ** 2
-    end.sum
-    stdev = variance / items.all.count
-    Math.sqrt(stdev).round(2)
   end
 
   def average_item_price_for_merchant(merchant_id)
@@ -66,18 +55,6 @@ class SalesAnalyst
       average_item_price_for_merchant(merchant.id)
     end
     (total / merchants.all.count).round(2)
-  end
-
-  def merchants
-    @sales_engine.merchants
-  end
-
-  def items
-    @sales_engine.items
-  end
-
-  def mean_calculation_merchant(merchant)
-    (items.find_all_by_merchant_id(merchant.id).count - average_items_per_merchant) ** 2
   end
 
   def calculate_stdev(mean)
